+++ Assets/_Project/Scripts/Infrastructure/ProjectInstaller.cs
@@ -1,5 +1,5 @@
-﻿using _Project.Infrastructure.InGameTime;
-using _Project.Infrastructure.SaveLoad;
+﻿using _Project.InGameTime;
+using _Project.SaveLoad;
 using _Project.Services.SceneLoader;
 using _Project.Services.AssetManagement;
 using _Project.Services.Factory;
