+++ Assets/_Project/Scripts/Services/PersistentProgress/_Project.PersistentProgress.asmdef
@@ -1,3 +1,14 @@
 {
-	"name": "_Project.PersistentProgress"
+    "name": "_Project.PersistentProgress",
+    "rootNamespace": "_Project.PersistentProgress",
+    "references": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
