+++ Assets/_Project/Scripts/Services/SaveLoad/SerializationDataExtensions.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace _Project.SaveLoad
+{
+    public static class SerializationDataExtensions
+    {
+        public static T ToDeserialized<T>(this string json) => JsonUtility.FromJson<T>(json);
+
+        public static string ToJson(this object obj) => JsonUtility.ToJson(obj);
+    }
+}
