+++ Assets/_Project/Scripts/Services/SaveLoad/_Project.SaveLoad.asmdef
@@ -1,7 +1,11 @@
 {
     "name": "_Project.SaveLoad",
-    "rootNamespace": "",
-    "references": [],
+    "rootNamespace": "_Project.SaveLoad",
+    "references": [
+        "GUID:161e9123cd2afcc4b9e5faa88d31cedd",
+        "GUID:6a66cbe93e5d50948a7153ed1e544fa0",
+        "GUID:3dc1fd8ce92a4bd4bba6e0d20b578142"
+    ],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
